move image_store fixture to conf.py
