+++ yt_idv/tests/test_yt_idv.py
@@ -1,12 +1,9 @@
 """Tests for `yt_idv` package."""
 
-import base64
-
 import numpy as np
 import pytest
 import yt
 import yt.testing
-from pytest_html import extras as html_extras
 
 import yt_idv
 from yt_idv import shader_objects
@@ -36,18 +33,6 @@ def osmesa_empty():
     rc.ds = ds
     yield rc
     rc.osmesa.OSMesaDestroyContext(rc.context)
-
-
-@pytest.fixture()
-def image_store(request, extras, tmpdir):
-    def _snap_image(rc):
-        image = rc.run()
-        img = yt.write_bitmap(image, None)
-        content = base64.b64encode(img).decode("ascii")
-        extras.append(html_extras.png(content))
-        extras.append(html_extras.html("<br clear='all'/>"))
-
-    return _snap_image
 
 
 def test_snapshots(osmesa_fake_amr, image_store):
